+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Enemy/Enemy.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-[CreateAssetMenu(fileName = "ENM_", menuName = "Touhou Spell Up/Danmaku/Enemy", order = 3)]
-public class Enemy : GameEntity
-{
-    // GameEntityの'property'フィールドをEnemyPropertyとしてキャストして公開する
-    public EnemyProperty PropertyTyped => property as EnemyProperty;
-}