+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Enemy/EnemyProperty.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-[CreateAssetMenu(fileName = "ENMP_", menuName = "Touhou Spell Up/Danmaku/EnemyProperty", order = 2)]
-public class EnemyProperty : GameEntityProperty
-{
-    [SerializeField] private int maxHealth = 100;
-    public int MaxHealth => maxHealth;
-}