+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Entity/Bullet/Bullet.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "BLT_", menuName = "Danmaku/Entity/Bullet/Bullet", order = 1)]
+public class Bullet : GameEntity
+{
+    // 親クラスのプロパティを隠蔽し、より具体的な型を返す新しいプロパティを定義します。
+    // これにより、シリアライズのエラーを回避しつつ、型安全性を維持します。
+    public new BulletProperty Property => base.Property as BulletProperty;
+}