+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Entity/Enemy/Enemy.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "ENM_", menuName = "Danmaku/Entity/Enemy/Enemy", order = 3)]
+public class Enemy : GameEntity
+{
+    // GameEntityの'property'フィールドをEnemyPropertyとしてキャストして公開する
+    public EnemyProperty PropertyTyped => property as EnemyProperty;
+}