+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Entity/Player/Player.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "PLY_", menuName = "Danmaku/Entity/Player/Player", order = 1)]
+public class Player : GameEntity
+{
+    public PlayerProperty PlayerProperty => property as PlayerProperty;
+    public override GameEntityProperty Property => PlayerProperty;
+}