+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Entity/Player/PlayerProperty.cs
@@ -1,0 +1,23 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "PLYP_", menuName = "Danmaku/Entity/Player/PlayerProperty", order = 0)]
+public class PlayerProperty : GameEntityProperty
+{
+    [Header("移動")]
+    [SerializeField] float moveSpeed = 12f;
+    [SerializeField] float moveSpeedSlow = 6f;
+    [SerializeField] Rect movableArea;
+
+    [Header("ショット")]
+    [SerializeField] float shotInterval = 0.15f;
+    [SerializeField] private PatternBase shotPatternNormal;
+    [SerializeField] private PatternBase shotPatternSlow;
+
+
+    public float MoveSpeed => moveSpeed;
+    public float MoveSpeedSlow => moveSpeedSlow;
+    public Rect MovableArea => movableArea;
+    public float ShotInterval => shotInterval;
+    public PatternBase ShotPatternNormal => shotPatternNormal;
+    public PatternBase ShotPatternSlow => shotPatternSlow;
+}